requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
docker
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
prometheus-api-client==0.5.4
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0
psutil==5.9.8
pycparser==2.21
PyMySQL==1.1.0
requests==2.31.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
APScheduler
kafka-python